Implements the five non-bulk transaction endpoints, and widens the amount
column ahead of them so the handlers are written against the real bound.
Bulk create/delete follows in #47.

Endpoints (inside the RequireAuth group, every query through db.WithUserTx):
- GET/POST /api/transactions
- GET/PATCH/DELETE /api/transactions/{id}

## The widening came first, deliberately

int4 milliunits capped a single transaction at 2,147,483.647 ARS — roughly
USD 1,400 at the BCRA reference rate, below a month's rent, a medical bill, or
a used-car payment. That is a product defect for an Argentine finance app, not
a theoretical limit. Implementing the handlers against a cap we already knew
was wrong would have meant writing throwaway int32 bounds checks and tests and
then deleting them.

It shipped as one change because any partial version leaves two caps in one
API. The jsonb_to_recordset column list in BulkCreateTransactions had to move
to bigint too: left at integer, bulk-create and the entire CSV import path
would have kept the old limit while single create no longer had it — and the
import path is the one users hit with a year of history.

The API bound is narrower than the column on purpose: ±(2^53-1) milliunits is
JavaScript's safe-integer limit, so every amount returned is exact in a
browser. Drizzle moved to bigint('amount', { mode: 'number' }), which is
required rather than cosmetic — node-postgres returns int8 as a string, so
without it the legacy stack would start receiving "12500".

## Behavior

Two projections that must not be merged: the list joins the current account
and category names, get/create/update return the transaction's own columns.
Renaming a category has to change the next list without rewriting any
transaction, which is why those names are joined data.

Dates resolve in UTC from a single captured now, with legacy's rules ported
exactly — including that a provided `to` does not re-anchor the default
`from`, and the 366/367-day boundary. Presence is distinguished from
emptiness: only omission defaults, because none of these parameters sets
allowEmptyValue and each references a schema with a minimum length.

Reference ownership is validated inside the same transaction as the write,
account before category, matching legacy's observable precedence. A foreign
key proves a row exists, not that the caller owns it, and FK checks bypass
RLS, so these lookups produce the 404s while RLS stays the backstop. The
unlocked window between check and write is mapped rather than left as a 500: a
concurrent delete surfaces as the same 404 the check would have returned.

Amounts are bounds-checked before any conversion. The contract states the
bound but oapi-codegen emits no range validation from minimum/maximum, so that
check is the only enforcement — without it a narrowing conversion wraps
silently and turns a large income into a negative expense.

Currency is required and must be exactly ARS; legacy has no currency field at
all, so nothing was broken by requiring it. categoryId treats omitted and null
identically as "no category", while an empty string fails validation as a
malformed id rather than silently clearing the association.

The mutation limiter is its own package: 60 per 60 seconds per (user, action),
mutex-guarded because Go serves requests concurrently where Node's event loop
serialized the original for free, with an injected clock so no test sleeps and
rejected attempts left unrecorded so a client cannot lock itself out.

## Review

Six findings across three rounds, all addressed and all threads resolved.

The sharpest was a doc comment claiming a concurrency guarantee the code did
not provide. Two findings were re-raised after I initially declined them, and
both were right: an explicitly empty query parameter is malformed under
allowEmptyValue's false default, and an empty categoryId belongs in request
validation rather than being reported as a missing reference. Tests that had
pinned the wrong behavior were inverted rather than deleted. The final finding
noted that adding a non-date use of InvalidDateQueryError left its generated
documentation lying; the component is now InvalidQueryError with examples for
both cases.

## Documentation

Ships with docs/api/transactions.md and docs/product/transactions.md — an
implementation reference covering what the contract cannot express, and a
product document covering the rules, limits, messages, and current gaps.
Issue #73 tracks extending that pairing to the rest of the app after #27.

Verification: full backend suite green against live PostgreSQL, go vet clean,
openapi:validate 28 operations, tsc --noEmit clean, bun lint clean, bun test
17/17. Registry entries 0014 and 0015 record the two behavior divergences;
0006 is closed out by the widening.

Closes #46.

+++ backend/internal/db/finance_schema_test.go
@@ -58,7 +58,10 @@ func TestFinanceBaseSchema_LiveDatabase(t *testing.T) {
 			table: "transactions",
 			columns: []columnSpec{
 				{"id", "text", false},
-				{"amount", "int4", false},
+				// int8 since migration 00005: int4 milliunits capped a single
+				// transaction near 2.15M ARS (~USD 1,400), which is below
+				// ordinary Argentine expenses like rent or a medical bill.
+				{"amount", "int8", false},
 				{"payee", "text", false},
 				{"notes", "text", true},
 				{"date", "timestamp", false},
